Ask Aider to do a rename of the file
This went well because I just told it to rename the file but didn't give it a new filename.

It did a great job of guessing the filename.

Aider still fails to run cli commands because it is running them in a subdirectory of the git repo. Not sure how to fix that.

In the mean time, I'm just using a terminal window to run the dev server and the build in terminal in Cursor to run things that might fail.

This way I can easily add the output to the chat and have Cursor help me troubleshoot.