Aider is pretty good at finding things that a developer would miss if you ask it about any other changes.
For example, I asked 'Are there any changes in the agents directory that need to be made knowing that the web acceptance tests now start the server before running?'
and it was able to make the README changes.

This step might help keep the README in sync because it always gets out of sync.