Change the default command.
Prompts:
Change the default command in the @inputHandler.ts to print out "Unknown command".

I no longer wish to keep the WeatherCommand. Can you delete all references but do not delete WeatherAgent.

Can you get rid of any unused code?

It changed the command but didn't figure out that it no longer needed the WeatherCommand.

Even asking to delete that didn't get rid of the function parameter.

This is a great example of a single file refactoring that needs a human in the loop to finish it.

I wonder if a different set of prompts would have made a difference?