Even with the new Claude 3.5 Sonnet model, refactorings can still be a back and forth, just like with pairing.
In the short term, any refactorings within a file should be done as a converstation between the humand and the AI.

The AI can start the process by using the prompt:

Can you suggest refactorings for @inputHandler.ts. Do not execute them, just come up with a plan.

Then you can either iterate on the plan or have it execute it and then iterate.

If you keep files small, then it should be easy for a human to parse the diff
or read the new file and propose changes.