Did the refactor into a module by hand.
TestUtils changed so that I could see the syntax error
in the test output.

Change the tsc config in the previous commit to make that a "compilation"
error for faster feedback.

The rest is making the commands a separate module.

Now that I know about modules, I understand why the llm kept wanting
to creates a types.ts file. That's just the way things go in JS/TS
module land.