Get the running of tests green
I wrote this in IntelliJ with Copilot with Claude turned on.

IntelliJ refactoring and navigation is much better than cursor
but I'm still learning how to efficiently use Copilot compared to Cursor.

For example, the ability to apply diffs to existing code from Cursor chat
is amazing!

Copilot you need to apply the whole thing and that's a pain.

May need to change to another IDE plugin because I am so much more familar
with IntellJ that VSCode/Cursor.

Will I be faster in one or the other?