This was a whirlwind of a coding session trying to get eveything in line.
I broke all kinds of rules but my goal wasn't production code, it
was to create a steel thread spike to get the layout of the
code into a better spot.

The next session will move forward from here with some tests and cleanup
to move this much closer to production code.