Get coding agent working with a local model.
This was a crazy one.

For the longest time, I couldn't get the system prompt right
to be able to get the model to return just the json.

When I asked Claude Sonnet in the browser to help me craft the
prompt to get the model to only return json, it worked flawlessly.

Lesson learned here is that promt engineering can make or break
getting your agent to do what you want.