More tweaks to get the AI to do what I want.
Observation:
Sometimes the assistant does what you ask it to do and sometimes
it doesn't.

These prompt tweaks are a effort to increase that chance.

This process feels like you are pairing with someone of mid to jr
level in their experience. Most of the time they get it right but
sometimes they forget.

It is also interesting to me to be made aware of all of the steps that
I have internalized in the software development process.

The process of tweaking the prompt feels like it is recording the evolution
of my process but on an accelerated timeframe because I'm having to
"educate" the ai assistant.