Remove diff instructions and add explanation steps
The diff rules don't work well in Windsurf so just get rid of them
for now.

Adding in some explanation steps to help the model reason through changes
it needs to make before proceeding. I'm hoping that this will help
prevent it from misreading the instructions from advent of code.